+++ index.js
@@ -1,5 +1,5 @@
 const Synth = require('./hydra-synth.js')
-//const ShaderGenerator = require('./shader-generator.js')
+const ShaderGenerator = require('./shader-generator.js')
 
 module.exports = Synth