+++ shader-generator.js
@@ -1,17 +1,14 @@
 const Generator = require('./src/generator-factory.js')
 const Sandbox = require('./src/eval-sandbox.js')
 
+const baseUniforms = ['s0', 's1', 's2', 's3', 'o0', 'o1', 'o2'] // names of uniforms usually used in hydra. These can be customized
+
 class ShaderGenerator {
-  constructor(opts) {
+  constructor({ defaultUniforms = {time: 0, resolution: [1280, 720]}, customUniforms = baseUniforms, extendTransforms = []} = {}) {
     var self = this
     self.renderer = {}
-    var defaultOpts = {
-      defaultUniforms: {
-        time: 0,
-        resolution: [1280, 720]
-      }
-    }
-    var generatorOpts = Object.assign({}, defaultOpts, opts)
+
+    var generatorOpts = { defaultUniforms, extendTransforms }
     generatorOpts.changeListener = ({type, method, synth}) => {
         if (type === 'add') {
           self.renderer[method] = synth.generators[method]
@@ -23,10 +20,16 @@ class ShaderGenerator {
     }
     this.generator = new Generator(generatorOpts)
     this.sandbox = new Sandbox(this.renderer, false)
+
+    this.initialCode = `
+      ${customUniforms.map((name) => `const ${name} = () => {}`).join(';')}
+    `
+    console.log(this.initialCode)
   }
 
   eval(code) {
-    this.sandbox.eval(code)
+    this.sandbox.eval(`${this.initialCode}
+          ${code}`)
     return this.generatedCode
   }
 }